Eliminated compiler warnings

+++ src/main/cxx/optimize.h
@@ -319,7 +319,7 @@ RQ::optimize(objective_function& f, double xw[], size_t n,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int k = 0; k < population_size; ++k) {
+      for (size_t k = 0; k < population_size; ++k) {
             fitness[k] = f(&x[k][0], n);
             index[k] = k;
         }
@@ -479,7 +479,7 @@ RQ::optimize(objectp obj, functionp f, double xw[], size_t n,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int k = 0; k < population_size; ++k) {
+        for (size_t k = 0; k < population_size; ++k) {
             fitness[k] = (obj->*f)(&x[k][0], n);
             index[k] = k;
         }
@@ -654,7 +654,7 @@ RQ::optimize(objective_function& f, double xw[], size_t n,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int k = 0; k < population_size; ++k) {
+        for (size_t k = 0; k < population_size; ++k) {
             fitness[k] = f(&x[k][0], n);
             index[k] = k;
         }
@@ -1000,7 +1000,7 @@ RQ::optimize(objective_function& f, double xw[], size_t n, constraint& reject,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int k = 0; k < population_size; ++k) {
+        for (size_t k = 0; k < population_size; ++k) {
             fitness[k] = f(&x[k][0], n);
             index[k] = k;
         }
@@ -1171,7 +1171,7 @@ RQ::optimize(objectp obj, functionp f, double xw[], size_t n, constraint& reject
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int k = 0; k < population_size; ++k) {
+        for (size_t k = 0; k < population_size; ++k) {
             fitness[k] = (obj->*f)(&x[k][0], n);
             index[k] = k;
         }
@@ -1287,7 +1287,7 @@ RQ::scale_cm(objective_function& f, const double x[], size_t n,
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (int g = 0; g < m; ++g) {
+        for (size_t g = 0; g < m; ++g) {
             valarray<double> q(x, n);
             for (size_t i = 0, j = g % n, ij = j; i < n; ++i, ij += n)
                 q[i] += s * z[g] * B[ij] * d[j];
